fix(banners): /admin/banners 500 — move pure filter out of the 'use client' module
The /admin/banners Server Component called getActiveSystemBanners, which was
exported from system-banner-source.ts ('use client'). A Server Component calling
a function exported from a 'use client' module receives a client-reference proxy
and throws at request time, so the page returned a stable 500.

next build does NOT catch this (it compiles fine; it's a runtime RSC boundary
error) — which is why CI stayed green. It was surfaced by a live route sweep.

Fix: extract the pure, server-safe filter (getActiveSystemBanners + isWithinWindow
+ matchesSurface + types) into a new non-client module
src/lib/banners/active-banners.ts. system-banner-source.ts keeps only the client
useActiveSystemBanners hook (delegating to the pure fn); the page imports
getActiveSystemBanners from the server-safe module.

eslint + typecheck clean.
